+++ test/CMakeLists.txt
@@ -8,5 +8,8 @@ target_link_libraries(module_lmdb PRIVATE LMDB spdlog ${LIBAMETSUCHI_NAME})
 
 AddTest(currency_test ametsuchi/currency_test.cc)
 target_link_libraries(currency_test PRIVATE ${LIBAMETSUCHI_NAME})
+
+AddTest(ametsuchi_test ametsuchi/ametsuchi.cc)
+target_link_libraries(ametsuchi_test PRIVATE ${LIBAMETSUCHI_NAME})
 #AddTest(module_sqlite module/sqlite_test.cc)
 #target_link_libraries(module_sqlite PRIVATE sqlite)
